remove sign-in page
